Rails 5.1 compat: coerce Relation arg to SQL string before matching against it
Makara hijacks `select_rows` and inspects the first argument, expecting
it to be a SQL string, to determine whether to route to primary or
replica connection.

In https://github.com/rails/rails/pull/25522 the method signature
changed to take an Arel relation instead of a SQL string, so all queries
will go to the primary database and none to the replica.

To fix, we'll need to either expect & duly inspect the Relation or call
`#to_sql` before checking. The first will take more code and possibly
introduce bugs; the second will mean a perf hit due to converting every
relation to SQL twice.

Fixes #149.