Remove context from connection pool management
It's always set to the current context when the pool is intialised, and
then compared against the current context as well. The only case in
which it could be relevant is when the context is forced to be a new one
manually, in the middle of a request. This will be handled in a
different way.