Use a "staging" area in context to avoid TTL from starting within request
That is, when we stick a proxy to master in a request and the request is
very long, if we set the expiration time based on the TTL right at the
moment of sticking, we might unstick before the request even ends. This
is probably not the intention, so instead of setting the expiration time
right when the proxy is stuck, we wait until the context is committed,
and at that time we move the newly stuck proxies to the storage area
from the staging area.